Moved to common folder to decouple from game
