+++ source/dagger/gameplay/common/cooldown_manager.h
